requirements.txt
google-auth==2.29.0
google-auth-httplib2==0.2.0
requests==2.32.3
google-analytics-data==0.18.9
google-cloud-bigquery==3.21.0
google-generativeai==0.7.2
jinja2==3.1.4
pytest==8.2.2
pytest-mock==3.14.0
